D-3 + L-1: consolidate CRS axis under pyramids.base.crs
D-3: osr.SpatialReference() hand-rolled in 6 modules
L-1: CRS axis scattered; no pyramids.base.crs
